need to fix form and write to database
