+++ client/src/components/SignIn/SignIn.css
@@ -1,5 +1,5 @@
 .App {
-  text-align: center;
+  width:200pxç;
 }
 
 .input-group{
